Title: 무~, Time: 125ms, Memory: 10MB, Status: Wrong Answer - CodeTree

+++ 241024/무~/moo.java
@@ -5,7 +5,7 @@ public class Main {
         // 여기에 코드를 작성해주세요.
         Scanner sc = new Scanner(System.in);
 
-        int n = sc.nextInt();
+        Long n = sc.nextLong();
         if (n%3 == 1) {
             System.out.print("m");
         } else {
